+++ core/src/integrationTest/java/tech/pegasys/web3signer/core/jsonrpcproxy/support/TestBaseConfig.java
@@ -123,11 +123,6 @@ public class TestBaseConfig implements BaseConfig {
   public void validateArgs() {}
 
   @Override
-  public Boolean isSwaggerUIEnabled() {
-    return false;
-  }
-
-  @Override
   public Boolean isAccessLogsEnabled() {
     return false;
   }