+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Upcoming Releases
+### Breaking Changes
+- `--swagger-ui-enabled` cli option has been removed. The Specs UI can be accessed at https://consensys.github.io/web3signer/.
+
+### Features Added
+
+
+### Bugs Fixed
+
+---
 ## 25.11.0
 ### Features Added
 - Updated Teku libraries to add scheduling for Fulu upgrade for mainnet
@@ -10,6 +20,7 @@ imported via the keymanager API to disk. This is *not* safe for operators who do
 - Updated web3j library to 5.0.1
 - Reduced Web3Signer docker image size. Removed unnecessary tools `iputils-ping, net-tools` from the image.
 
+---
 ## 25.9.1
 ### Breaking changes
 - `eea*` RPC methods and other privacy supporting classes have been removed since Privacy feature has been removed from [Besu 25.6.0](https://github.com/hyperledger/besu/releases/tag/25.6.0)